+++ paimon-web-plugin/paimon-web-plugin-cryptography/paimon-web-plugin-cryptography-none/src/main/java/org/apache/paimon/web/cryptography/NoneSymmetricCryptoService.java
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.paimon.web.cryptography;
+
+import org.apache.paimon.web.cyptography.SymmetricCryptoService;
+
+/** NoneSymmetricCryptoService does not perform any operations on the incoming data. */
+public class NoneSymmetricCryptoService implements SymmetricCryptoService {
+
+    @Override
+    public String encrypt(String data) {
+        return data;
+    }
+
+    @Override
+    public String decrypt(String data) {
+        return data;
+    }
+}
